+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.12.0 (Nov 11, 2025)
+* Added support for reading logs in Nullstone.
+* Added support for displaying metrics in Nullstone.
+* Optimized aggregation of capabilities.
+
 # 0.11.0 (Sep 22, 2025)
 * Upgrade terraform providers.